refactor: port server methods to node
