+++ crates/edr_rpc_server/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 async-rwlock = { version = "1.3.0", default-features = false }
 async-trait = { version = "0.1.73", default-features = false }
 axum = "0.6.18"
+indexmap = { version = "2.0.0", default-features = false, features = ["std"] }
 hex = { version = "0.4.3", default-features = false, features = ["alloc"] }
 hyper = { version = "0.14.27", default-features = false, features = ["server"] }
 k256 = { version = "0.13.1", default-features = false, features = ["arithmetic", "ecdsa", "pkcs8", "precomputed-tables", "std"] }
@@ -31,3 +32,9 @@ tracing-subscriber = { version = "0.3.17", default-features = false, features = 
 cargo_toml = { version = "0.15.3", default-features = false }
 serde = { version = "1.0.147", default-features = false, features = ["derive"] }
 toml = { version = "0.5.9", default-features = false }
+
+[features]
+## This feature needs to be accessible from integration tests. If it's not default, we'd have to run the integration
+## tests `cargo test --features test-tools`
+default = ["test-tools"]
+test-tools = []
